refactor: atualiza rota, modelo e repositórios relacionados a Cliente e UF
- Altera rota no ClientesController para `api/v1/cliente`
- Adiciona método `CarregarTodasCidades` com tratamento de erros aprimorado
- Modifica chave primária de `UF` de `long` para `int`
- Atualiza modelos `Cidade`, `UF` e `RastreamentoPedidosContextModelSnapshot` para refletir nova tipagem de `idUF`
- Refatora `ClienteRepositoryDapper` e `UFRepository` para uso de Dapper e novas assinaturas de métodos
- Atualiza `IUFRepository` para aceitar `int` em `CarregarPorId` e adiciona `CarregarTodasUf`
- Cria classe `UfQueries` para centralizar consultas SQL de UF

+++ RastreamentoPedidos/Model/Clientes/Cidade.cs
@@ -8,7 +8,7 @@ namespace RastreamentoPedidos.Model.Clientes
         //[Key]
         public long idCidade { get; set; }
         public string nome { get; set; } = string.Empty;
-        public long idUF { get; set; }
+        public int idUF { get; set; }
         public UF UF { get; set; } = new UF();
     }
 }
